Rename Image to mockImage for consistency
